Handle abandoned & overrun errors
If varnish is restarted, vago.Log() busy loops endlessly.

VSLQ_Dispatch() can return 2 more error conditions that were not
handled, abandoned-or-closed and overrun. In such cases there is nothing
to do be done but to reinitiate the VSM. Thus, we return the error to
the caller that can safely create a new Varnish struct.